Add human-readable identifiers to messages
This new feature adds human-readable `id`s on patterns.  Previously,
an identifier of `0` was patched on the first pattern.  Now, `her-him`
is used.

This identifier is additionally patched on messages as a `ruleId`.
This allows for additional documentation to go more in depth regarding
why certain phrases are highlighted.

In addition, a new `source` property is patched on message too,
which signals that a message originates from `retext-equality`.